fix: Fix capturing functor in lambda
Capture were always made by copy, which was breaking a test. That's now fixed.
However fixing that broke the free-function systems, as a local reference was created in the register function, which was then treated as an lvalue reference. We're now moving pointer in these function to prevent that.

Also fix a test where the wrong parameter was sent, which meant that a instruction that was expected to throw didn't