fix: Callback event could not be dispatched through a std::any
cause: dispatcher::_dispatch_callback re-computed the type_index.
This cause an issue with std::any as parameter, since it used std::any
instead of the actual event type for the index.

solution: dispatcher::dispatch was modified to properly compute the
type_index for types inside a std::any, which removed the need for an
overload. dispatcher::_dispatch_callback now takes the index as a
parameter, since it's already computed by dispatcher::dispatch